[LA][APB-282] Refactor out some common helper code.
